requirements.txt
Flask==3.0.2
psycopg==3.1.18
pytest==8.1.1
pytest-playwright==0.4.4
pytest-xprocess==0.23.0
py==1.11.0
Flask-Login==0.6.3